fixed flatpickr date issue
